test(checkout): CHECKOUT-9131 convert orderConfirmation enzyme tests to rtl
